+++ TEST/sponsorCoinEthers_JS/index.html
@@ -1,0 +1,78 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <link rel="stylesheet" href="./styles.css" />
+  <meta charset="UTF-8" />
+  <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>SponsorCoin.js</title>
+</head>
+
+<body>
+  <h3>Sponsor Coin</h3>
+  <table>
+    <tbody id="connectWallet_TBL">
+      <tr>
+        <th>Wallet Calls</th>
+        <th class="addr">Parameters</th>
+        <td class="spacer"></td>
+        <th>Test Account</th>
+        <th class="addr">Test Address</th>
+      </tr>
+      <tr>
+        <td>
+          <button id="connectWallet_BTN" class="cell_BTN"
+            onclick="GUI_connectWallet(this.id, document.getElementById('connectWallet_TX').value)">
+            Connect Wallet
+          </button>
+        </td>
+        <td><input type="text" id="connectWallet_TX" value="METAMASK" /></td>
+        <td class="spacer"></td>
+        <td>SponsorCoin Contract</td>
+        <td>0x334710ABc2Efcc3DF2AfdA839bF8d0dA923dB36A</td>
+      </tr>
+      <tr>
+        <th>Account Calls</th>
+        <td class="addr"></td>
+        <td class="spacer"></td>
+        <td>Meta Mask Account 1</td>
+        <td>0x4F75f07232a56c2b98FC9878F496bFc32e317Ace</td>
+      </tr>
+      <tr>
+        <td>
+          <button id="activeAccount_BTN" class="cell_BTN" onclick="GU_getActiveAccount(this.id)">
+            Active Account
+          </button>
+        </td>
+        <td><input type="text" id="activeAccount_TX" /></td>
+        <td class="spacer"></td>
+        <td>Meta Mask Ledger 1</td>
+        <td>0xFAbed8e2f3a29aEE5002087F1140Ef4C6ACa25B4</td>
+      </tr>
+      <tr>
+        <td>
+          <button id="ethereumAccountBalance_BTN" class="cell_BTN" onclick="GUI_getEthereumAccountBalance(this.id)">
+            Ethereum Account Balance
+          </button>
+        </td>
+        <td><input type="text" id="ethereumAccountBalance_TX" /></td>
+        <td class="spacer"></td>
+        <td>Meta Mask Ledger 2</td>
+        <td>0x1EFFDE4A0e5eEcF79810Dd39f954A515ab962D63</td>
+      </tr>
+    </tbody>
+  </table>
+
+  <button onclick="document.location.reload(true);">Reload Page</button>
+
+  <script type="module">
+    import { ethers } from "https://cdn.ethers.io/lib/ethers-5.2.esm.min.js";
+    // Your code here...
+</script>
+    <script src="utils/utils.js" type="application/javascript"></script>
+    <script src="wallet/wallet.js" type="application/javascript"></script>
+    <script src="wallet/WEB_Wallet.js" type="application/javascript"></script>
+</body>
+
+</html>
