+++ TEST/sponsorCoinEthers_JS/wallet/WEB_Wallet.js
