+++ TEST/sponsorCoinEthers_JS/wallet/wallet.js
