--- V2/Ethers JS Metamask Connection/index.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Learn Ethers.js</title>
-    
-</head>
-<body>
-    <button onclick="connectMetamask()">connect</button>
-    <button onclick="getBalance()">getBalance</button>
-    <button onclick="readDataFromSmartContract()">readDataFromSmartContract</button>
-    <br/>
-    <button onclick="sendUsdtToAccount()">sendUsdtToAccount</button>
-    <button onclick="deployContract()">deployContract</button>
-    <br/>
-
-    <button onclick="incrementNumberOnSmartContract()">incrementNumberOnSmartContract</button>
-    <button onclick="emitAnEvent()">emitAnEvent</button>
-    <button onclick="listenToEvents()">listenToEvents</button>
-    <br/>
-    <button onclick="sendEtherWhenCallingFunction()">sendEtherWhenCallingFunction</button>
-
-    <button onclick="readFromSCOnctractStorage()">readFromSCOnctractStorage</button>
-
-    <script src="https://cdn.ethers.io/lib/ethers-5.2.umd.min.js"
-        type="application/javascript"></script>
-        <script src="index.js"></script>
-    
-</body>
-</html>
